[onert] Write ChromeTracingObserver to three files by EventWriter
- Introduce new class EventWriter which takes EventRecorder and writes
its Records to files
- Now EventRecorder is a kind of data class to be recorded

ONE-DCO-1.0-Signed-off-by: Yongseop Kim <yons.kim@samsung.com>
